// src/app/interface/form-value-interface.ts
export interface FormValueInterface {
  S: string;
  distance: number;
  e: string;
  group: string;
  mProt: number;
  mTran: number;
  oDep: number;
  oRec: number;
  pRt: number;
  tEs: number;
  tIns: number;
  team: string;
  try: string;
  x: string;
  y: string;
  z: string;
  penalty: string;
  timeStamp?: number;
  id?: string;
}